fixed edit post bug

+++ controllers/feed.js
@@ -4,6 +4,7 @@ const path = require('path');
 const { validationResult } = require('express-validator/check');
 
 const Post = require('../models/post');
+const User = require('../models/user');
 
 exports.getPosts = (req, res, next) => {
   const currentPage = req.query.page || 1;
@@ -17,13 +18,11 @@ exports.getPosts = (req, res, next) => {
         .skip((currentPage - 1) * perPage)
         .limit(perPage)
         .then((posts) => {
-          res
-            .status(200)
-            .json({
-              message: 'Posts fetched successfully',
-              posts: posts,
-              totalItems: totalItems,
-            });
+          res.status(200).json({
+            message: 'Posts fetched successfully',
+            posts: posts,
+            totalItems: totalItems,
+          });
         })
         .catch((err) => {
           if (!err.statusCode) {
@@ -52,21 +51,28 @@ exports.createPost = (req, res, next) => {
   // console.log(imageUrl);
   const title = req.body.title;
   const content = req.body.content;
+  let creator;
   const post = new Post({
     title: title,
     content: content,
     imageUrl: imageUrl,
-    creator: {
-      name: 'Nemanja',
-    },
+    creator: req.userId,
   });
   post
     .save()
     .then((result) => {
-      // console.log(result);
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      creator = user;
+      user.posts.push(post);
+      return user.save();
+    })
+    .then((result) => {
       res.status(201).json({
         message: 'Post created successfully!',
-        post: result,
+        post: post,
+        creator: { _id: creator._id, name: creator.name },
       });
     })
     .catch((err) => {
@@ -82,6 +88,7 @@ exports.createPost = (req, res, next) => {
 exports.getPost = (req, res, next) => {
   const postId = req.params.postId;
   Post.findById(postId)
+    .populate('creator')
     .then((post) => {
       if (!post) {
         const error = new Error('Could not find post.');
@@ -108,7 +115,7 @@ exports.updatePost = (req, res, next) => {
   }
   const title = req.body.title;
   const content = req.body.content;
-  let imageUrl = req.file.path.replace('\\', '/');
+  let imageUrl = req.body.image;
 
   if (req.file) {
     imageUrl = req.file.path.replace('\\', '/');
@@ -125,19 +132,26 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized!');
+        error.statusCode = 403;
+        throw error;
+      }
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl);
       }
       post.title = title;
-      post.content = content;
       post.imageUrl = imageUrl;
+      post.content = content;
       return post.save();
     })
     .then((result) => {
-      // console.log('post updated');
+      console.log(result);
       res.status(200).json({ message: 'Post updated!', post: result });
     })
     .catch((err) => {
+      // console.log(err);
+
       if (!err.statusCode) {
         err.statusCode = 500;
       }
@@ -155,6 +169,11 @@ exports.deletePost = (req, res, next) => {
         throw error;
       }
       // Check logged in user
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized!');
+        error.statusCode = 403;
+        throw error;
+      }
       clearImage(post.imageUrl);
       return Post.findByIdAndDelete(postId);
     })
